+++ src/clients/infra/Assets.ts
@@ -2,7 +2,7 @@ import { contains, filter, isEmpty, pick as ramdaPick, zipObj } from 'ramda'
 import { Readable } from 'stream'
 
 import { AppMetaInfo } from '../..'
-import { CacheType, inflightURL, InstanceOptions } from '../../HttpClient'
+import { CacheType, inflightURL, InstanceOptions, RequestTracingConfig } from '../../HttpClient'
 import {
   IgnoreNotFoundRequestConfig,
 } from '../../HttpClient/middlewares/notFound'
@@ -54,7 +54,7 @@ export class Assets extends InfraClient {
     this.routes = createRoutes(this.context.workspace)
   }
 
-  public getSettings (dependencies: AppMetaInfo[], appAtMajor: string, params: AssetsParams = {}) {
+  public getSettings (dependencies: AppMetaInfo[], appAtMajor: string, params: AssetsParams = {}, tracingConfig?: RequestTracingConfig) {
     const {pick, files} = params
     const filtered = this.getFilteredDependencies(appAtMajor, dependencies)
 
@@ -63,49 +63,49 @@ export class Assets extends InfraClient {
       const buildJson = useBuildJson(dependency, appVendorName)
 
       return buildJson
-        ? this.getBuildJSONForApp(dependency, appVendorName, pick)
-        : this.getSettingsFromFilesForApp(dependency, files)
+        ? this.getBuildJSONForApp(dependency, appVendorName, pick, tracingConfig)
+        : this.getSettingsFromFilesForApp(dependency, files, tracingConfig)
     }
     ))
   }
 
-  public async getBuildJSONForApp(app: AppMetaInfo, appVendorName: string, pick: string | string[] = []): Promise<Record<string, any>> {
+  public async getBuildJSONForApp(app: AppMetaInfo, appVendorName: string, pick: string | string[] = [], tracingConfig?: RequestTracingConfig): Promise<Record<string, any>> {
     const pickArray = Array.isArray(pick) ? pick : [pick]
-    const buildJson: Record<string, any> = await this.getJSON(app.id, `dist/${appVendorName}/build.json`)
+    const buildJson: Record<string, any> = await this.getJSON(app.id, `dist/${appVendorName}/build.json`, undefined, tracingConfig)
     const result = !isEmpty(pickArray) ? ramdaPick(pickArray, buildJson) : buildJson
 
     result.declarer = app.id
     return result
   }
 
-  public async getSettingsFromFilesForApp(app: AppMetaInfo, files: string | string[] = []): Promise<Record<string, any>> {
+  public async getSettingsFromFilesForApp(app: AppMetaInfo, files: string | string[] = [], tracingConfig?: RequestTracingConfig): Promise<Record<string, any>> {
     // If there's no support for build.json, then fetch individual files and zip them into an {[file]: content} object.
     const filesArray = Array.isArray(files) ? files : [files]
-    const fetched = await Promise.all(filesArray.map(file => this.getJSON(app.id, file, true)))
+    const fetched = await Promise.all(filesArray.map(file => this.getJSON(app.id, file, true, tracingConfig)))
     const result: Record<string, any> = zipObj(filesArray, fetched)
 
     result.declarer = app.id
     return result
   }
 
-  public async getJSON<T extends object | null>(appId: string, file: string, nullIfNotFound?: boolean) {
+  public async getJSON<T extends object | null>(appId: string, file: string, nullIfNotFound?: boolean, tracingConfig?: RequestTracingConfig) {
     const locator = parseAppId(appId)
     const linked = !!locator.build
 
     if (linked) {
-      return this.getAppJSONByAccount<T>(appId, file, nullIfNotFound)
+      return this.getAppJSONByAccount<T>(appId, file, nullIfNotFound, tracingConfig)
     }
-    return this.getAppJSONByVendor<T>(appId, file, nullIfNotFound)
+    return this.getAppJSONByVendor<T>(appId, file, nullIfNotFound, tracingConfig)
   }
 
-  public async getFile(appId: string, file: string, nullIfNotFound?: boolean) {
+  public async getFile(appId: string, file: string, nullIfNotFound?: boolean, tracingConfig?: RequestTracingConfig) {
     const locator = parseAppId(appId)
     const linked = !!locator.build
 
     if (linked) {
-      return this.getAppFileByAccount(appId, file, nullIfNotFound)
+      return this.getAppFileByAccount(appId, file, nullIfNotFound, tracingConfig)
     }
-    return this.getAppFileByVendor(appId, file, nullIfNotFound)
+    return this.getAppFileByVendor(appId, file, nullIfNotFound, tracingConfig)
   }
 
   public getFilteredDependencies(apps: string | string[], dependencies: AppMetaInfo[]): AppMetaInfo[] {
@@ -114,7 +114,7 @@ export class Assets extends InfraClient {
     return filter(depends, dependencies)
   }
 
-  public getAppBundleByVendor = (app: string, bundlePath: string, generatePackageJson: boolean): Promise<Readable> => {
+  public getAppBundleByVendor = (app: string, bundlePath: string, generatePackageJson: boolean, tracingConfig?: RequestTracingConfig): Promise<Readable> => {
     const locator = parseAppId(app)
     const params = generatePackageJson && {_packageJSONEngine: 'npm', _packageJSONFilter: 'vtex.render-builder@x'}
     const metric = locator.build ? 'apps-get-bundle' : 'registry-get-bundle'
@@ -125,52 +125,76 @@ export class Assets extends InfraClient {
       },
       metric,
       params,
+      tracing: {
+        requestSpanNameSuffix: metric,
+        ...tracingConfig?.tracing,
+      },
     })
   }
 
-  protected getAppJSONByAccount = <T extends object | null>(app: string, path: string, nullIfNotFound?: boolean) => {
+  protected getAppJSONByAccount = <T extends object | null>(app: string, path: string, nullIfNotFound?: boolean, tracingConfig?: RequestTracingConfig) => {
     const locator = parseAppId(app)
     const inflightKey = inflightURL
+    const metric = 'assets-get-json-by-account'
     return this.http.get<T>(this.routes.Files(this.context.account, locator, path), {
       cacheable: CacheType.Memory,
       inflightKey,
-      metric: 'assets-get-json-by-account',
+      metric,
       nullIfNotFound,
+      tracing: {
+        requestSpanNameSuffix: metric,
+        ...tracingConfig?.tracing,
+      },
     } as IgnoreNotFoundRequestConfig)
   }
 
-  protected getAppJSONByVendor = <T extends object | null>(app: string, path: string, nullIfNotFound?: boolean) => {
+  protected getAppJSONByVendor = <T extends object | null>(app: string, path: string, nullIfNotFound?: boolean, tracingConfig?: RequestTracingConfig) => {
     const locator = parseAppId(app)
     const vendor = locator.name.split('.')[0]
     const inflightKey = inflightURL
+    const metric = 'assets-get-json-by-vendor'
     return this.http.get<T>(this.routes.Files(vendor, locator, path), {
         cacheable: CacheType.Any,
         inflightKey,
-        metric: 'assets-get-json-by-vendor',
+        metric,
         nullIfNotFound,
+        tracing: {
+          requestSpanNameSuffix: metric,
+          ...tracingConfig?.tracing,
+        },
       } as IgnoreNotFoundRequestConfig)
   }
 
-  protected getAppFileByAccount = (app: string, path: string, nullIfNotFound?: boolean) => {
+  protected getAppFileByAccount = (app: string, path: string, nullIfNotFound?: boolean, tracingConfig?: RequestTracingConfig) => {
     const locator = parseAppId(app)
     const inflightKey = inflightURL
+    const metric = 'assets-get-file-by-account'
     return this.http.getBuffer(this.routes.Files(this.context.account, locator, path), {
       cacheable: CacheType.Any,
       inflightKey,
-      metric: 'assets-get-file-by-account',
+      metric,
       nullIfNotFound,
+      tracing: {
+        requestSpanNameSuffix: metric,
+        ...tracingConfig?.tracing,
+      },
     })
   }
 
-  protected getAppFileByVendor = (app: string, path: string, nullIfNotFound?: boolean) => {
+  protected getAppFileByVendor = (app: string, path: string, nullIfNotFound?: boolean, tracingConfig?: RequestTracingConfig) => {
     const locator = parseAppId(app)
     const vendor = locator.name.split('.')[0]
     const inflightKey = inflightURL
+    const metric = 'assets-get-file-by-vendor'
     return this.http.getBuffer(this.routes.Files(vendor, locator, path), {
       cacheable: CacheType.Any,
       inflightKey,
-      metric: 'assets-get-file-by-vendor',
+      metric,
       nullIfNotFound,
+      tracing: {
+        requestSpanNameSuffix: metric,
+        ...tracingConfig?.tracing,
+      },
     })
   }
 }