+++ src/clients/infra/BillingMetrics.ts
@@ -1,4 +1,4 @@
-import { InstanceOptions } from '../../HttpClient'
+import { InstanceOptions, RequestTracingConfig } from '../../HttpClient'
 import { IOContext } from '../../service/worker/runtime/typings'
 import { InfraClient } from './InfraClient'
 
@@ -7,8 +7,12 @@ export class BillingMetrics extends InfraClient {
     super('colossus@0.x', context, options)
   }
 
-  public sendMetric = (metric: BillingMetric) =>
-    this.http.post<BillingMetric>('/metrics', metric)
+  public sendMetric = (metric: BillingMetric, tracingConfig?: RequestTracingConfig) => {
+    return this.http.post<BillingMetric>('/metrics', metric, { tracing: {
+      requestSpanNameSuffix: 'send-metric',
+      ...tracingConfig?.tracing,
+    }})
+  }
 }
 
 export interface BillingMetric {