+++ src/tracing/spanReference/SpanReferenceTypes.ts
@@ -1,0 +1,4 @@
+export enum SpanReferenceTypes {
+    CHILD_OF = 0,
+    FOLLOWS_FROM = 1,
+}